feat(tallies): conversion from tallies to Array2View and to Numpy
